tests: Add tests for split-dirs
Add tests for the situation where the source file and the coverage
file reside in directories that are not subdirs of each other.  This is
a common situation for projects built with autotools and meson.

Add a first test on cov--coverage, also using split-dir.

Add failing testcase cov--get-buffer-coverage-test-split-dir-relative
that exposes the problem reported in #23.